Revert "add create user"
