--- backend/src/auth/auth.controller.ts
@@ -1,14 +1,0 @@
-import { Controller, Post, UseGuards, Request } from '@nestjs/common';
-import { AuthService } from './auth.service';
-import { LocalAuthGuard } from './guard/local-auth.guard';
-
-@Controller('auth')
-export class AuthController {
-  constructor(private readonly authService: AuthService) {}
-
-  @UseGuards(LocalAuthGuard)
-  @Post('login')
-  async login(@Request() req) {
-    return req.user;
-  }
-}
