--- backend/src/auth/auth.module.ts
@@ -1,11 +1,0 @@
-import { Module } from '@nestjs/common';
-import { PassportModule } from '@nestjs/passport';
-import { UserModule } from '../user/user.module';
-import { AuthService } from './auth.service';
-import { LocalStrategy } from './strategies/local.strategy';
-
-@Module({
-  imports: [UserModule, PassportModule],
-  providers: [AuthService, LocalStrategy],
-})
-export class AuthModule {}
