--- backend/src/files/files.entity.ts
@@ -1,13 +1,0 @@
-import { Column, Entity, PrimaryGeneratedColumn } from 'typeorm';
-import { ApiProperty } from '@nestjs/swagger';
-
-@Entity()
-export class Files {
-  @ApiProperty({ type: Number })
-  @PrimaryGeneratedColumn()
-  id: number;
-
-  @ApiProperty({ type: String })
-  @Column({ type: 'varchar', nullable: true })
-  imageUrl?: string;
-}
