--- backend/src/user/dtos/user-update.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/swagger';
-import { UserCreateDto } from './user-create.dto';
-
-export class UserUpdateDto extends PartialType(UserCreateDto) {}
