--- backend/src/user/user.controller.ts
@@ -1,26 +1,0 @@
-import { Body, Controller, Get, Post, UploadedFile, UseInterceptors } from '@nestjs/common';
-import { FileInterceptor } from '@nestjs/platform-express';
-import { UserService } from './user.service';
-import { UserCreateDto } from './dtos/user-create.dto';
-import { User } from './user.entity';
-
-@Controller('user')
-export class UserController {
-  constructor(private readonly userService: UserService) {}
-
-  @Get()
-  findAll(): Promise<User[]> {
-    return this.userService.findAll();
-  }
-
-  @Post()
-  create(@Body() createUser: UserCreateDto) {
-    return this.userService.create(createUser);
-  }
-
-  @Post('avatar')
-  @UseInterceptors(FileInterceptor('image'))
-  async uploadAvatar(@UploadedFile() avatar: Express.Multer.File) {
-    return this.userService.updateAvatar(avatar);
-  }
-}
